Removed all other queries except cartesian join, tested with client, reading db values from database.properties file
