+++ build.gradle
@@ -1,0 +1,26 @@
+group 'com.rds.dbwrapper'
+
+apply plugin: 'java'
+apply plugin: 'war'
+
+repositories {
+    mavenLocal()
+    mavenCentral()
+}
+
+dependencies {
+
+    compile 'org.json:json:20140107'
+    compile 'org.glassfish.jersey.core:jersey-server:2.5.1'
+    compile 'org.slf4j:slf4j-api:1.7.6'
+    compile 'org.slf4j:slf4j-log4j12:1.7.5'
+    compile 'com.fasterxml.jackson.jaxrs:jackson-jaxrs-json-provider:2.2.3'
+    compile 'org.glassfish.jersey.containers:jersey-container-servlet:2.5.1'
+    compile 'mysql:mysql-connector-java:5.1.18'
+
+    compile files('libs/ojdbc6.jar')
+}
+
+war {
+    baseName="rds-dbwrapper";
+}
