+++ src/main/resources/database.properties
@@ -1,0 +1,9 @@
+oracleDriver = oracle.jdbc.OracleDriver
+oracleUrl = jdbc:oracle:thin:@//localdocker:1521/xe
+oracleUsername = APPDY
+oraclePassword = APPDY
+
+mysqlDriver = com.mysql.jdbc.Driver
+mysqlUrl = jdbc\:mysql\://db\:3306/appdy?useUnicode\=true&characterEncoding\=UTF-8&autoReconnect\=true
+mysqlUsername = root
+mysqlPassword = tiger
